filesSize zu bigint ändern für Dateien > 4GB

+++ sql/filemanager.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS `fmg_files` (
   `files_folderID` int(10) unsigned NOT NULL,
   `files_userID` int(10) unsigned NOT NULL,
   `filesDate` int(10) unsigned NOT NULL,
-  `filesSize` int(10) unsigned NOT NULL,
+  `filesSize` bigint(20) unsigned NOT NULL,
   `filesThumbnail` tinyint(3) unsigned NOT NULL,
   PRIMARY KEY (`filesID`)
 ) ENGINE=InnoDB  DEFAULT CHARSET=utf8 AUTO_INCREMENT=1 ;
